Refactor smart playlist composables and sort translations
- Split useSmartPlaylistRulesForm into orchestrator with 4 sub-composables
- Add useSmartPlaylistSeedItems, useSmartPlaylistGenres, useSmartPlaylistContentFilters, useSmartPlaylistYearRange
- Sort smart_playlist translation keys alphabetically in en.json
