+++ src/components/smart_playlist/constants.ts
@@ -1,0 +1,1 @@
+export const SMART_PLAYLIST_PROVIDER_DOMAIN = "smart_playlist";
