+++ src/translations/en.json
@@ -1631,48 +1631,48 @@
     "radiobrowser_by_popularity": "By popularity",
     "radiobrowser_by_category": "By category",
     "restore_genre": "Restore",
-    "genre_name": "Genre name",
     "smart_playlist": {
+        "any": "Any",
+        "clear": "Clear",
         "create": "Create Smart Playlist",
+        "created": "Smart playlist created!",
+        "creating": "Creating…",
+        "dedup_hours": "No repeat within",
+        "dedup_hours_tooltip": "Skip tracks that were recently played. Set a window in hours (1–168) during which a track will not be repeated. Leave at 0 to disable.",
         "desc": "Generate a playlist from your library based on rules.",
-        "mode": "Mode",
         "dynamic": "Dynamic",
-        "fixed": "Fixed (one-time)",
-        "track_count": "Number of tracks",
+        "edit_rules": "Edit rules",
+        "excluded_albums": "Exclude albums",
+        "excluded_artists": "Exclude artists",
+        "excluded_genres": "Exclude genres",
         "favorites_only": "Favorites only",
+        "fixed": "Fixed (one-time)",
+        "limit": "Max tracks",
         "logic": "Rule combination",
         "logic_and": "Match ALL rules (AND)",
         "logic_or": "Match ANY rule (OR)",
         "logic_tooltip": "AND: a track must match all active rules. OR: a track needs to match at least one rule.",
         "min_popularity": "Minimum popularity",
-        "any": "Any",
-        "limit": "Max tracks",
+        "min_popularity_tooltip": "Popularity is a score from 0–100 provided by your streaming service. Tracks without a popularity score are always included.",
+        "mode": "Mode",
+        "off": "Off",
+        "saving": "Saving…",
+        "section_advanced": "Advanced",
+        "section_content": "Artists & Albums",
+        "section_filters": "Filters",
+        "section_source": "Source",
+        "seed_artist": "Similar to artist",
+        "seed_artist_library_only": "Only use tracks from my library",
+        "seed_artist_pick": "Pick an artist",
+        "seed_artist_tooltip": "Pick an artist to find similar artists and their tracks via your streaming provider. Artist and album filters are ignored when a seed artist is set.",
+        "seed_overrides_filter": "Not used when a seed track or artist is set",
         "seed_track": "Similar to track",
         "seed_track_pick": "Pick a track",
         "seed_track_tooltip": "Pick a track to find similar tracks via your streaming provider (e.g. Apple Music, Spotify). Artist and album filters are ignored when a seed track is set.",
-        "seed_artist": "Similar to artist",
-        "seed_artist_pick": "Pick an artist",
-        "seed_artist_tooltip": "Pick an artist to find similar artists and their tracks via your streaming provider. Artist and album filters are ignored when a seed artist is set.",
-        "seed_artist_library_only": "Only use tracks from my library",
-        "seed_overrides_filter": "Not used when a seed track or artist is set",
-        "excluded_artists": "Exclude artists",
-        "excluded_albums": "Exclude albums",
-        "excluded_genres": "Exclude genres",
-        "dedup_hours": "No repeat within",
-        "dedup_hours_tooltip": "Skip tracks that were recently played. Set a window in hours (1–168) during which a track will not be repeated. Leave at 0 to disable.",
-        "off": "Off",
-        "min_popularity_tooltip": "Popularity is a score from 0–100 provided by your streaming service. Tracks without a popularity score are always included.",
-        "clear": "Clear",
-        "creating": "Creating…",
-        "created": "Smart playlist created!",
-        "edit_rules": "Edit rules",
-        "saving": "Saving…",
-        "year_range": "Year range",
+        "track_count": "Number of tracks",
         "year_from": "From year",
-        "year_to": "To year",
-        "section_source": "Source",
-        "section_content": "Artists & Albums",
-        "section_filters": "Filters",
-        "section_advanced": "Advanced"
-    }
+        "year_range": "Year range",
+        "year_to": "To year"
+    },
+    "genre_name": "Genre name",
 }
